// src/generated/model/userGoogleRequest.ts
/**
 * Datafood Server API
 * API contract of Datafood server application BRASIDAS
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleModel } from './roleModel';


export interface UserGoogleRequest { 
    /**
     * email of the user
     */
    email?: string;
    /**
     * username of the user
     */
    username?: string;
    /**
     * image url of the user
     */
    image?: string;
    /**
     * google id of the user
     */
    idGoogle?: string;
    /**
     * family name of the user
     */
    familyName?: string;
    /**
     * given name of the user
     */
    givenName?: string;
    /**
     * boolean whether email is verified or not
     */
    emailVerified?: boolean;
    /**
     * username of the user
     */
    sessionExpiration?: number;
    /**
     * user roles
     */
    roles?: Array<RoleModel>;
}